requirements.txt
pandas>=1.3.0
numpy>=1.21.0
matplotlib>=3.4.0
plotly>=5.0.0
yfinance>=0.1.63
pandas-ta>=0.3.14
scipy>=1.7.0
PyYAML>=5.4.0
pytest>=6.2.0
requests>=2.25.0
